+++ src/Host/Fuse.Cli/Host/Rpc/FuseHostJsonContext.cs
@@ -21,4 +21,6 @@ namespace Fuse.Cli.Rpc;
 [JsonSerializable(typeof(ScopeFileDto))]
 [JsonSerializable(typeof(DiagnosticsDto))]
 [JsonSerializable(typeof(SecretDiagnosticDto))]
+[JsonSerializable(typeof(ExplainResultDto))]
+[JsonSerializable(typeof(ExplainFileDto))]
 public sealed partial class FuseHostJsonContext : JsonSerializerContext;